requirements.txt
ultralytics
opencv-python
torch
websockets
asyncio
numpy
